Completed assignment six
Finished assignment six and feel pretty confident about the solution I
made for the assignment. Will check with the solutions folder after this
commit to see if I am close.

Learned that if I wanted to remove bullet points from lists, I should use
list-style-type: none; and to remove the underline from links use text-decoration: none;.
Also learned that even though text-decoration: none; is put on the parent level,
it will not affect the children. This is not 100% confirmed but will see if
the solutions also did it the way I did.