Modified assignment six
I was right about putting the text-decoration: none; into the parent
will not do a thing but also could have changed the list items directly instead
of changing them with classes. Also forgot about the margin and padding
on item lists (ul{}) are a bit strange and need to change them to 0. Did not
need to justify-content: space-between; for body class, instead should have used
flex: 1; to separate the header, content, and footer. Some reason there was a
margin-bottom: 16px; placed in the input{} selector so had to delete that, but was not
present in the solution.css file. Everything else was fine.